pyproject.toml
[project]
name = "mysql-hub-mcp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "aiohttp>=3.12.15",
    "google-adk>=1.10.0",
    "litellm>=1.75.4",
    "rich>=14.1.0",
]

[tool.uv.workspace]
members = [
    "client",
    "server",
    "bridge",
    "agent",
]
